+++ app/products/[slug]/_components/ProductImageGallery.tsx
