+++ app/products/[slug]/_components/ProductInfo/ProductCategories.tsx
