+++ app/products/[slug]/_components/RelatedProducts.tsx
