+++ tsconfig.json
@@ -21,6 +21,7 @@
     "baseUrl": ".",
     "paths": {
       "~/*": [
+        "./app/*",
         "./src/*"
       ]
     },
